.ipynb_checkpoints/requirements-checkpoint.txt
tensorflow>=2.10.0
streamlit>=1.22.0
pillow>=9.5.0
numpy>=1.23.5
matplotlib>=3.5.0
opencv-python>=4.5.5 